Fixed a lot of things.

+++ src/main/java/dao/TweetDAO.java
@@ -7,7 +7,7 @@ public interface TweetDAO {
 
     List<Tweet> findAll();
 
-    Tweet findById(int id);
+    Tweet findById(long id);
 
     List<Tweet> findByMessage(String query);
